Move rewind of inner batch 0 file
For non-zero batches, inner batch file is rewound at the end of
ExecHashJoinNewBatch(). This is part of the preparation setup for the
next batch. However, when batch 0 spills, this is too late to rewind the
batch as there is no batch previous batch to call setup. Hence the
fence-post-problem of setting up batch 0.